arm: SMP: Fix RCU stalls due to mis-backported upstream patch
Samsung backported the aforementioned patch, available here:
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux.git;a=commit;h=7deabca0acfe02b8e18f59a4c95676012f49a304

However, 3.0.x has an additional ipi_timer() that needs to be wrapped
with irq_enter()/irq_exit(), this patch adds these.

Signed-off-by: Andrew Dodd <atd7@cornell.edu>

arm: smp: Wrap backtrace IPI with irq_enter/exit

This appears to be something that should also be wrapped.  Also,
the backtraces experienced may be a cause and not a symptom -
instead of an RCU stall causing backtraces, backtraces
may be causing an RCU stall - which causes more backtraces
and then the system gets mental.

Change-Id: If1c24afd497dd77b0505ad6caf74216b068eed36
Signed-off-by: Andrew Dodd <atd7@cornell.edu>